requirements.txt
Django>=4.2
django-telegram-bot
python-dotenv
